+++ docs/changes.rst
@@ -7,6 +7,8 @@ Weblate 5.15.2
 
 .. rubric:: Improvements
 
+* :ref:`addon-weblate.generate.generate` is now triggered upon installation.
+
 .. rubric:: Bug fixes
 
 .. rubric:: Compatibility